Move more regions (snake_case)
    renamed:    src/nupic/regions/KNNClassifierRegion.py ->
src/nupic/regions/knn_classifier_region.py
    renamed:    src/nupic/regions/PluggableEncoderSensor.py ->
src/nupic/regions/pluggable_encoder_sensor.py
    renamed:    src/nupic/regions/SDRClassifierRegion.py ->
src/nupic/regions/sdr_classifier_region.py
