+++ docs/source/quick-start/network.rst
@@ -194,7 +194,7 @@ classifier region. In our case, we are interested in:
     probabilities = classifierRegion.getOutputData("probabilities")
 
 
-Refer to the documentation of :class:`~nupic.regions.SDRClassifierRegion.SDRClassifierRegion` for
+Refer to the documentation of :class:`~nupic.regions.sdr_classifier_region.SDRClassifierRegion` for
 more information about output values and their structure.
 
 We'll use the helper function below to extract predictions more easily from
