+++ src/nupic/engine/__init__.py
@@ -46,11 +46,11 @@ pyRegions = (
     ("nupic.regions.anomaly_region", "AnomalyRegion"),
     ("nupic.regions.cla_classifier_region", "CLAClassifierRegion"),
     ("nupic.regions.knn_anomaly_classifier_region", "KNNAnomalyClassifierRegion"),
-    ("nupic.regions.KNNClassifierRegion", "KNNClassifierRegion"),
-    ("nupic.regions.PluggableEncoderSensor", "PluggableEncoderSensor"),
+    ("nupic.regions.knn_classifier_region", "KNNClassifierRegion"),
+    ("nupic.regions.pluggable_encoder_sensor", "PluggableEncoderSensor"),
     ("nupic.regions.PyRegion", "PyRegion"),
     ("nupic.regions.record_sensor", "RecordSensor"),
-    ("nupic.regions.SDRClassifierRegion", "SDRClassifierRegion"),
+    ("nupic.regions.sdr_classifier_region", "SDRClassifierRegion"),
     ("nupic.regions.SPRegion", "SPRegion"),
     ("nupic.regions.SVMClassifierNode", "SVMClassifierNode"),
     ("nupic.regions.TestRegion", "TestRegion"),
