+++ src/nupic/regions/knn_anomaly_classifier_region.py
@@ -29,7 +29,7 @@ import numpy
 
 from nupic.algorithms import anomaly
 from nupic.bindings.regions.PyRegion import PyRegion
-from KNNClassifierRegion import KNNClassifierRegion
+from knn_classifier_region import KNNClassifierRegion
 from nupic.bindings.math import Random
 from nupic.frameworks.opf.exceptions import (HTMPredictionModelInvalidRangeError,
                                              HTMPredictionModelInvalidArgument)
