+++ src/nupic/regions/knn_classifier_region.py
