+++ src/nupic/regions/pluggable_encoder_sensor.py
