+++ src/nupic/regions/sdr_classifier_region.py
